backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
google-generativeai
groq
python-multipart
python-dotenv
joblib
uuid
